fix(api-stream): Handle body correctly before throwing error
