+++ lib/api/api-stream.js
@@ -4,12 +4,11 @@ const { finished } = require('stream')
 const {
   InvalidArgumentError,
   InvalidReturnValueError,
-  RequestAbortedError
+  RequestAbortedError, ResponseStatusCodeError
 } = require('../core/errors')
 const util = require('../core/util')
 const { AsyncResource } = require('async_hooks')
 const { addSignal, removeSignal } = require('./abort-signal')
-const Readable = require('./readable')
 
 class StreamHandler extends AsyncResource {
   constructor (opts, factory, callback) {
@@ -79,7 +78,7 @@ class StreamHandler extends AsyncResource {
   }
 
   onHeaders (statusCode, rawHeaders, resume, statusMessage) {
-    const { factory, opaque, context, abort, callback } = this
+    const { factory, opaque, context } = this
 
     if (statusCode < 200) {
       if (this.onInfo) {
@@ -89,22 +88,8 @@ class StreamHandler extends AsyncResource {
       return
     }
 
-    const parsedHeaders = util.parseHeaders(rawHeaders)
-    const contentType = parsedHeaders['content-type']
-
     this.factory = null
     const headers = this.responseHeaders === 'raw' ? util.parseRawHeaders(rawHeaders) : util.parseHeaders(rawHeaders)
-
-    if (this.throwOnError && statusCode >= 400) {
-      const body = new Readable(resume, abort, contentType)
-      this.res = body
-
-      this.runInAsyncScope(util.getResolveErrorBodyCallback, null,
-        { callback, body, contentType, statusCode, statusMessage, headers }
-      )
-      return
-    }
-
     const res = this.runInAsyncScope(factory, null, {
       statusCode,
       headers,
@@ -132,6 +117,12 @@ class StreamHandler extends AsyncResource {
       }
 
       this.callback = null
+
+      if (this.throwOnError && statusCode >= 400) {
+        this.runInAsyncScope(callback, null, new ResponseStatusCodeError(`Response status code ${statusCode}${statusMessage ? `: ${statusMessage}` : ''}`, statusCode, headers))
+        return
+      }
+
       this.runInAsyncScope(callback, null, err || null, { opaque, trailers })
 
       if (err) {
