+++ lib/core/util.js
@@ -5,7 +5,7 @@ const { kDestroyed, kBodyUsed } = require('./symbols')
 const { IncomingMessage } = require('http')
 const stream = require('stream')
 const net = require('net')
-const { InvalidArgumentError, ResponseStatusCodeError } = require('./errors')
+const { InvalidArgumentError } = require('./errors')
 const { Blob } = require('buffer')
 const nodeUtil = require('util')
 const { stringify } = require('querystring')
@@ -361,33 +361,6 @@ function isFormDataLike (chunk) {
 const kEnumerableProperty = Object.create(null)
 kEnumerableProperty.enumerable = true
 
-async function getResolveErrorBodyCallback ({ callback, body, contentType, statusCode, statusMessage, headers }) {
-  if (statusCode === 204 || !contentType) {
-    body.dump()
-    process.nextTick(callback, new ResponseStatusCodeError(`Response status code ${statusCode}${statusMessage ? `: ${statusMessage}` : ''}`, statusCode, headers))
-    return
-  }
-
-  try {
-    if (contentType.startsWith('application/json')) {
-      const payload = await body.json()
-      process.nextTick(callback, new ResponseStatusCodeError(`Response status code ${statusCode}${statusMessage ? `: ${statusMessage}` : ''}`, statusCode, headers, payload))
-      return
-    }
-
-    if (contentType.startsWith('text/')) {
-      const payload = await body.text()
-      process.nextTick(callback, new ResponseStatusCodeError(`Response status code ${statusCode}${statusMessage ? `: ${statusMessage}` : ''}`, statusCode, headers, payload))
-      return
-    }
-  } catch (err) {
-    // Process in a fallback if error
-  }
-
-  body.dump()
-  process.nextTick(callback, new ResponseStatusCodeError(`Response status code ${statusCode}${statusMessage ? `: ${statusMessage}` : ''}`, statusCode, headers))
-}
-
 module.exports = {
   kEnumerableProperty,
   nop,
@@ -415,6 +388,5 @@ module.exports = {
   validateHandler,
   getSocketInfo,
   isFormDataLike,
-  buildURL,
-  getResolveErrorBodyCallback
+  buildURL
 }
